+++ eql/__init__.py
@@ -68,7 +68,7 @@ from .walkers import (
     Walker,
 )
 
-__version__ = '0.8.6'
+__version__ = '0.8.7'
 __all__ = (
     "__version__",
     "AnalyticOutput",